Video_models/Density_models/OMAN/requirements.txt
albumentations==1.3.0
geomloss==0.2.6
matplotlib==3.7.1
numpy==1.24.3
opencv_python==4.7.0.72
opencv_python_headless==4.7.0.72
Pillow==9.4.0
scipy==1.13.1
timm==0.9.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.4
gdown